requirements.txt
lightgbm>=4.3.0
numpy>=1.24.0
pandas>=2.1.0
pyarrow>=14.0.0
requests>=2.31.0
scikit-learn>=1.3.0
shap>=0.44.0
tqdm>=4.66.0
pyyaml>=6.0.0
rich>=13.7.0
joblib>=1.3.0
